+++ uniffi_bindgen/src/bindings/python/templates/Types.py
@@ -67,9 +67,6 @@
 {%- when Type::Record { name, module_path } %}
 {%- include "RecordTemplate.py" %}
 
-{%- when Type::Object { name, module_path, imp } %}
-{%- include "ObjectTemplate.py" %}
-
 {%- when Type::Timestamp %}
 {%- include "TimestampHelper.py" %}
 
@@ -91,9 +88,21 @@
 {%- when Type::Custom { name, module_path, builtin } %}
 {%- include "CustomType.py" %}
 
-{%- when Type::External { name, module_path, namespace, kind, tagged } %}
+{%- when Type::External { name, module_path, namespace, .. } %}
 {%- include "ExternalTemplate.py" %}
 
+{%- else %}
+{%- endmatch %}
+{%- endfor %}
+
+# objects.
+{%- for type_ in self.iter_sorted_object_types() %}
+{%- match type_ %}
+{%- when Type::Object { name, .. } %}
+{%-     let type_name = type_|type_name %}
+{%-     let ffi_converter_name = type_|ffi_converter_name %}
+{%-     let canonical_type_name = type_|canonical_name %}
+{%-     include "ObjectTemplate.py" %}
 {%- else %}
 {%- endmatch %}
 {%- endfor %}