+++ uniffi_bindgen/src/interface/function.rs
@@ -270,6 +270,9 @@ pub trait Callable {
         }
     }
 
+    // Scaffolding function
+    fn ffi_func(&self) -> &FfiFunction;
+
     // Quick way to get the rust future scaffolding function that corresponds to our return type.
 
     fn ffi_rust_future_poll(&self, ci: &ComponentInterface) -> String {
@@ -313,9 +316,13 @@ impl Callable for Function {
     fn is_async(&self) -> bool {
         self.is_async
     }
+
+    fn ffi_func(&self) -> &FfiFunction {
+        &self.ffi_func
+    }
 }
 
-// Needed because Askama likes to add extra refs to variables
+// Needed because Rinja likes to add extra refs to variables
 impl<T: Callable> Callable for &T {
     fn arguments(&self) -> Vec<&Argument> {
         (*self).arguments()
@@ -331,6 +338,10 @@ impl<T: Callable> Callable for &T {
 
     fn is_async(&self) -> bool {
         (*self).is_async()
+    }
+
+    fn ffi_func(&self) -> &FfiFunction {
+        (*self).ffi_func()
     }
 
     fn takes_self(&self) -> bool {